- Replaced Emu de MIDI main player core with my own stream parser - Version is now 1.143
git-tfs-id: [http://localhost:8080/tfs/DefaultCollection/]$/foobar2000/files/plugins.root;C726

+++ emidi_alpha/DsaCommon.hpp
@@ -15,10 +15,10 @@ typedef unsigned short WORD;
 typedef unsigned long  DWORD;
 typedef unsigned char  UINT8;
 typedef unsigned short UINT16;
-typedef unsigned long  UINT32;
+typedef unsigned int  UINT32;
 typedef signed char  INT8;
 typedef signed short INT16;
-typedef signed long  INT32;
+typedef signed int  INT32;
 
 typedef unsigned int UINT;
 typedef signed int INT;
